Refactor: python, cts: Move StonithdTest into the _cts module.
No code changes.
