+++ python/pacemaker/_cts/tests/__init__.py
@@ -12,4 +12,5 @@ from pacemaker._cts.tests.remotedriver import RemoteDriver
 from pacemaker._cts.tests.simulstartlite import SimulStartLite
 from pacemaker._cts.tests.simulstoplite import SimulStopLite
 from pacemaker._cts.tests.starttest import StartTest
+from pacemaker._cts.tests.stonithdtest import StonithdTest
 from pacemaker._cts.tests.stoptest import StopTest
